Use one tool for reading items instead of a tool per collection
Having a single tool per collections can cause issues with MCP clients
that have hard limits of number of tools available (i.e Cursor),
with Directus installs with many collections.
Additionally, when there are a lot of junction tables for many_to_many
relationships, the collection names can grow too long and the
tool-per-collection name will exceed the max length for a tool name.

This update creates a single tool to read items that takes a collection
name as param